refactor: change folder name to PricingPageComponents
