+++ wealth-sync/src/components/PricingPageComponents/NoPricingSection/FairPricingCard.tsx
