+++ wealth-sync/src/components/PricingPageComponents/NoPricingSection/NoPricingSection.tsx
